+++ src/main/java/chess/domain/position/Position.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.position;
 
 import java.util.Objects;
 import java.util.stream.IntStream;
