+++ src/main/java/chess/domain/strategy/MoveStrategy.java
@@ -1,7 +1,7 @@
 package chess.domain.strategy;
 
-import chess.domain.Direction;
-import chess.domain.Position;
+import chess.domain.position.Direction;
+import chess.domain.position.Position;
 import java.util.Deque;
 import java.util.Map;
