+++ src/main/java/chess/domain/strategy/speical/QueenMoveStrategy.java
@@ -1,6 +1,6 @@
-package chess.domain.strategy;
+package chess.domain.strategy.speical;
 
-import chess.domain.Direction;
+import chess.domain.position.Direction;
 import java.util.List;
 
 public class QueenMoveStrategy extends SpecialPieceMoveStrategy{
