+++ src/main/java/chess/domain/strategy/speical/RookMoveStrategy.java
@@ -1,6 +1,6 @@
-package chess.domain.strategy;
+package chess.domain.strategy.speical;
 
-import chess.domain.Direction;
+import chess.domain.position.Direction;
 import java.util.List;
 
 public class RookMoveStrategy extends SpecialPieceMoveStrategy{
