+++ src/main/java/chess/domain/strategy/speical/SpecialPieceMoveStrategy.java
@@ -1,7 +1,8 @@
-package chess.domain.strategy;
+package chess.domain.strategy.speical;
 
-import chess.domain.Direction;
-import chess.domain.Position;
+import chess.domain.position.Direction;
+import chess.domain.position.Position;
+import chess.domain.strategy.MoveStrategy;
 import java.util.ArrayDeque;
 import java.util.Deque;
 import java.util.List;
