+++ src/test/java/chess/domain/PositionTest.java
@@ -1,5 +1,8 @@
 package chess.domain;
 
+import chess.domain.position.Column;
+import chess.domain.position.Position;
+import chess.domain.position.Row;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
