+++ src/test/java/chess/domain/chessGame/ChessGameTest.java
@@ -3,11 +3,11 @@ package chess.domain.chessGame;
 import chess.domain.Board;
 import chess.domain.ChessGame;
 import chess.domain.Color;
-import chess.domain.Column;
+import chess.domain.position.Column;
 import chess.domain.Piece;
 import chess.domain.PieceType;
-import chess.domain.Position;
-import chess.domain.Row;
+import chess.domain.position.Position;
+import chess.domain.position.Row;
 import java.util.List;
 import java.util.Map;
 import org.assertj.core.api.Assertions;
